Fix macOS wheel zoom and bundle version
Preserve fractional ImGui mouse wheel input through the native, SWIG, and C# GUIManager bindings so slow mouse or trackpad scrolling still zooms the viewport.

Update the macOS app bundle version to 1.80.5 and sync packaged Info.plist versions from Core.Version during macOS packaging to avoid stale bundle metadata.

+++ Dev/Cpp/Viewer/GUI/efk.GUIManager.cpp
@@ -727,7 +727,7 @@ int GUIManager::GetMouseButton(int32_t mouseButton)
 	return window->GetMouseButton(mouseButton);
 }
 
-int GUIManager::GetMouseWheel()
+float GUIManager::GetMouseWheel()
 {
 	return ImGui::GetIO().MouseWheel;
 }
